+++ releasenotes.md
@@ -44,6 +44,7 @@
 * Player messages are now slightly translucent again when the `r_hud_translucency` CVAR is `on` and the `vid_widescreen` CVAR is `off`, but not when vanilla mode is enabled.
 * SSAA (supersampling anti-aliasing) is now applied to the help screen’s background when the <kbd>F1</kbd> key is pressed and the `r_supersampling` CVAR is `on`.
 * The design of the 3 digit used in the status bar when the `r_detail` CVAR is `high` has been tweaked slightly.
+* The use of the player’s fists and chainsaw are now displayed by the `playerstats` CCMD.
 
 ![](https://github.com/bradharding/www.doomretro.com/raw/master/wiki/bigdivider.png)